add authorization header for pass jwt token

+++ front-end/online-electoral-system/src/api/CandidateApiService.js
@@ -2,23 +2,23 @@ import { apiClient } from "./VoterApiService";
 
 //get all the candidate for a specific
 export const retrieveAllCandidateApi = (positionId) => {
-  return apiClient.get(`api/candidates/${positionId}`);
+  return apiClient.get(`api/v1/admin/candidates/${positionId}`);
 };
 
 export const retrieveCandidateApi = (id) => {
-  return apiClient.get(`api/candidates/id/${id}`);
+  return apiClient.get(`api/v1/admin/candidates/id/${id}`);
 }
 
 //post a candidate
 export const createCandidateApi = (candidate) => {
-  return apiClient.post("/api/candidates", candidate);
+  return apiClient.post("/api/v1/admin/candidates", candidate);
 };
 
 //delete a candidate
 export const deleteCandidateApi = (id) => {
-  return apiClient.delete(`/api/candidates/${id}`);
+  return apiClient.delete(`/api/v1/admin/candidates/${id}`);
 };
 
 export const updateCandidateApi = (id,candidate) => {
-  return apiClient.put(`/api/candidates/${id}`,candidate);
+  return apiClient.put(`/api/v1/admin/candidates/${id}`,candidate);
 }
